requirements.txt
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
simplejwt==2.0.1
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.2
